hw/omap3_boot.c: don't override image header detection for MMC images
Remove code which was erroneously overriding the detection of a CHSETUP
header in an OMAP boot image loaded from an MMC card. Newer versions
of x-loader create MLO files which include this CHSETUP header, and
if we don't correctly skip over it to the actual code then they won't
boot.

Signed-off-by: Peter Maydell <peter.maydell@linaro.org>