tracing: Add zlib based compression of "simple" trace files.
This patch adds a new "zsimple" trace backend which uses zlib to directly write
compressed trace files. The uncompressed trace file format of the "zsimple"
backend is identical to the "simple" backend.

Compressed tracefiles can be processed using the simpletrace.py script
directly:

> zcat trace.gz | scripts/simpletrace.py trace-event /dev/stdin

The "zsimple" backend is intended to be used with trace events which generate
huge amounts of trace information, like for exmaple, the "arm_idecode"
trace event introduced earlier in this patch series.