ARM TrustZone: Enable handling of IRQ and FIQ exceptions in secure monitor mode.
On TrustZone systems the secure configuration register (SCR) allows
IRQ and FIQ exceptions to be redirected to exception handlers in secure
monitor mode. This patch enables the corresponding functionality in
the emulation.

Support for dealing with (external) prefetch and instruction aborts (for
example caused by TrustZone memory protection violations) is currently
incomplete.