ARM TrustZone: Add infrastructure for TrustZone memory protection.
There is a broad range of possible mechanisms for partitioning the
memory and I/O space of a TrustZone enabled system into secure and
non-secure areas. This patch introduces a prototype framework intended
to cover a broad range of possible confguration mechanism.

The framework introduced in this patch models different protection regions
as hierarchy of virtual address space control busses (ASCBus) which connect
to virtual partition devices (ASCPartition) on top of the qdev device model.

Each "partition devices" is defined by a physical address range and two
bit masks defining the page access permissions for secure and non-secure
world.

This patch exposes the address space control subsystem in QEMU's device
tree by means of a virtual root address space controller sysbus device. This
sysbus device does not occupy any actual memory or IRQ resources. It only creates
a logical binding between QEMU's system bus and the virtual address space controller bus.

In addition to the memory protection framework this patch adds a device
model for the BP147 TrustZone Protection Controller. The BP147 peripherals provides a
series of decode protection registers which are used to configure the security domains
to which platform peripherals belong to. The secure RAM size register of th BP147
device is currently not fully implemented in this peripheral model.

Finally, this patch provides a skeleton implementation of the TrustZone Address Space
controller (TZC380). The skeleton implementation of the TZC380 exposes
all registers (except the "itcrg/itip/itop" test registers) of the peripheral.
A backend connection to the TrustZone infrastructure is not yet implemented.