arm: Allow semihosting to work on boards which do not provide an arm-bootinfo record.
This patch allows semihosting to be used (again) on board emulations (like
e.g. the BeagleBoard) which do not use the standard boot information via
the target-arm/arm_boot.c mechanism. Without this patch attempts to use
semihosting on such boards causes a blunt SIGSEGV when an application tries
to do a SYS_GET_CMDLINE semi-call.