target-arm: add helper to detect the processor security status.
This patch introduces a new arm_is_secure helper function for detecting
the current security status of an ARM TrustZone enabled core.

The CPU state is considered to be secure when at least one of the
following criteria is true:
* arm_feature reports a core _without_ TrustZone support
* The NS bit (bit 0) in the Secure Configuration Register (SCR) is cleared
* The core is in secure monitor mode and mon_is_secure is non-zero

Note that the ARM architecture treats secure monitor mode as secure
by definition, regardless of the status of the SCR.NS bit. Under special
circumstances it is necessary to access non-secure parts of the processor
state, like non-secure versions of banked co-processor registers, from within
secure monitor mode. For this purpose the mon_is_secure parameter of
arm_is_secure can be set to zero, in order to treat secure monitor mode
like any other processor mode (with respect to SCR.NS).