hw/omap3_boot.c: Account for root directory size when reading FAT16
In a FAT16 image, the first cluster comes straight after the root
directory, and clusters are numbered starting at two.

The code used to assume that the root directory was always exactly the
same size as two clusters.  This is only true on a hard-disk-style
image with 512 root-directory entries and 8k clusters (typically disk
images sized between 256M and 512M).

Signed-off-by: Peter Chubb <peter.chubb@nicta.com.au>